requirements.txt
streamlit>=1.30
requests>=2.31
fpdf2>=2.7
python-docx>=1.1
pdfplumber>=0.11
pdf2image>=1.17
pytesseract>=0.3
Pillow>=10.0
openai>=1.30
python-dotenv>=1.0
pandas>=2.0
